Added the requested changes
