+++ package.json
@@ -24,8 +24,7 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "server": "json-server --watch data.json --port 8000"
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
