+++ src/App.js
@@ -5,22 +5,38 @@ import { Container } from '@mui/material';
 import Heading from './components/Heading';
 import Paper from '@mui/material/Paper'
 import BottomNavigation from '@mui/material/BottomNavigation';
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
+import AddForm from './components/AddForm';
 
 function App() {
-  const [notes, setNotes] = useState([])
-  useEffect(() => {
-    const getNotes = async () => {
-      const notesFromServer = await fetchNotes()
-      setNotes(notesFromServer)
+  const [notes, setNotes] = useState([
+    {
+      id: 1,
+      title: "Definition of Done",
+      content: "Each task you have here should be made with the following procedure: - For each ticket, you should create a separate branch from the master branch"
+    },
+    {
+      id: 2,
+      title: "Layout of page",
+      content: "List of hardcoded notes (each note should have a title and short content), - Add button,- Edit button for each note,- Delete button for each note"
+    },
+    {
+      id: 3,
+      title: "Create note",
+      content: "- Subject/Title of Note ( required field ), - The content of note limited to 300 characters (required field)."
+    },
+    {
+      id: 4,
+      title: "Edit note",
+      content: "When a user clicks on edit, the same modal as Add note should be displayed with the same rules with few changes - Instead of Add Button, there should be an Edit button."
+    },
+    {
+      id: 5,
+      title: "Delete note",
+      content: "When a user clicks on delete button, confirmation modal should be displayed with a question Delete note title* ?"
     }
-    getNotes()
-  }, [])
-  const fetchNotes = async () => {
-    const res = await fetch('http://localhost:8000/notes')
-    const data = await res.json()
-    return data
-  }
+  ])
+  const [open, setOpen] = useState(false);
   const addNote = (note) => {
     const id=Math.floor(Math.random()*10000) + 1
     const newNote = {id, ...note}
@@ -32,8 +48,15 @@ function App() {
         sx={{ position:'fixed',bottom:0,left:0,right:0 }} 
         elevation={3}>
         <BottomNavigation>
-          <CustomizedDialogs onAdd={addNote}/>
-        </BottomNavigation >
+          <CustomizedDialogs 
+            open={open} 
+            setOpen = {setOpen}>
+            <AddForm 
+              onAdd = {addNote} 
+              open={open} 
+              setOpen = {setOpen}/>
+          </CustomizedDialogs>
+        </BottomNavigation > 
       </Paper>  
       <Heading />
       <Container sx={{marginTop:5, marginBottom:12}}>
