+++ src/components/Modal.js
@@ -8,13 +8,6 @@ import DialogTitle from '@mui/material/DialogTitle';
 import DialogContent from '@mui/material/DialogContent';
 import IconButton from '@mui/material/IconButton';
 import CloseIcon from '@mui/icons-material/Close';
-import Typography from '@mui/material/Typography';
-import TextField from '@mui/material/TextField';
-import { useState } from 'react';
-import Box from '@mui/material/Box';
-import Stack from '@mui/material/Stack';
-import AddCircleIcon from '@mui/icons-material/AddCircle';
-import CancelIcon from '@mui/icons-material/Cancel';
 import NoteAddIcon from '@mui/icons-material/NoteAdd';
 
 const BootstrapDialog = styled(Dialog)(({ theme }) => ({
@@ -52,38 +45,19 @@ BootstrapDialogTitle.propTypes = {
   onClose: PropTypes.func.isRequired,
 };
 
-export default function CustomizedDialogs({ onAdd }) {
-  const [open, setOpen] = React.useState(false);
-  const [titleNote, setTitleNote] = useState("");
-  const [contentNote, setContentNote] = useState("");
+export default function CustomizedDialogs({children,open,setOpen}) {
   const handleClickOpen = () => {
     setOpen(true);
   };
   const handleClose = () => {
     setOpen(false);
-    setTitle('');
-    setContent('');
-  };
-  const handleTitleChange = (event) => {
-    setTitleNote(event.target.value);
-  };
-  const handleContentChange = (event) =>{
-    setContentNote(event.target.value);
-  };
-  const [title, setTitle] = useState('');
-  const [content, setContent] = useState('');
-  const onSubmit = (e) => {
-    e.preventDefault()
-    onAdd({title,content})
-    setTitle('')
-    setContent('')
-  }
+  };  
   return (
     <Grid>
       <Button 
         sx={{backgroundColor:'#21b5b4',color:'white',width:200,height:57}} 
         variant="contained" 
-        onClick={handleClickOpen} 
+        onClick={handleClickOpen}
         startIcon={<NoteAddIcon />}>
           Add New Note
       </Button>
@@ -96,62 +70,7 @@ export default function CustomizedDialogs({ onAdd }) {
           Add new note
         </BootstrapDialogTitle>
         <DialogContent dividers>
-          <form onSubmit={onSubmit}>  
-            <Typography variant='paragraph' component='p' textAlign='center'>
-              <label>Add a new note subject/title</label> <br />
-            </Typography>
-            <TextField 
-              sx={{marginTop:1,backgroundColor:'#6ffbff',width:226}}
-              value={title}
-              onChange = {(e) => setTitle(e.target.value)}
-              required
-              name='title'
-              id="title"
-              label="Required"
-            />
-            <Box sx={{marginTop:2}}>
-              <Typography 
-                variant='paragraph' 
-                component='p' 
-                textAlign='center'>
-              <Box>
-                <label>Add content of note</label> <br />
-              </Box>
-            <TextField 
-              sx={{marginTop:1, backgroundColor:'#6ffbff', width:226}}
-              required
-              value={content}
-              onChange = {(e) => setContent(e.target.value)}
-              name='content'
-              id="content"
-              label="Required"
-              multiline
-              rows={4}
-              inputProps = {{maxLength: 300}}
-            />
-              </Typography><br />
-            </Box>
-            <Stack direction="row" spacing={2}>
-              <Button 
-                sx={{width:103, height:40, backgroundColor:'#0097b3'}} 
-                variant="contained"  
-                disabled={!title||!content} 
-                startIcon={<AddCircleIcon />} 
-                type='submit' 
-                autoFocus>
-                  Add
-              </Button>
-              <Button 
-                sx={{width:103, height:40, backgroundColor:'#0097b3'}} 
-                variant="contained" 
-                onClick={handleClose} 
-                startIcon={<CancelIcon />} 
-                type='reset' 
-                autoFocus>
-                  Cancel
-              </Button>
-            </Stack>
-          </form>
+          {children}
         </DialogContent>
       </BootstrapDialog>
     </Grid>
